Merge pull request #24 from Gary Kennedy.
Added the analytic implied volatility approximation of J. Choi, K Kim
and M. Kwak (2009), “Numerical Approximation of the Implied Volatility
Under Arithmetic Brownian Motion”, Applied Math. Finance, 16(3), pp.
261-268.

Given the premium of an option, it provides an analytic approximation of
implied volatility of the Bachelier formula. In interest rate the
volatility is referred to as bpvol.